Split project data class back into a separate file
